chore: Move some utility methods to submodules of scalar_funcs
